+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 -   Removed sdlgfx dependency, replacing them with a cache-based `Surface` implementation.
 -   Default z-index for all queue-based `Draw` functions is now 0 to match the rest of rubato. (was Math.INF)
 -   All `Draw` functions can now take in a camera as an optional argument.
+-   `Text` component now uses a Surface in its implementation.
 
 ### Removed